Restructured file folders and updated README
