+++ README.md
@@ -1,3 +1,82 @@
 # shiny-internal-rag-project
 
-This is from the Lonely Octopus program.
+![GitHub License](https://img.shields.io/badge/license-MIT-default.svg)
+
+## Description
+
+The primary goal of this project is to leverage AI technologies and provide innovation by seamlessly summarizing, retrieving, and performing key concept extraction on various LixCap documentation. This initiative will create a competitive advantage against industry peers, increase operational efficiency for the project management and research teams, and result in informed decision making across the company.
+
+## Table of Contents
+
+* [Installation](#installation)
+* [Usage](#usage)
+* [Video](#video)
+* [Deployed Site](#deployed-site)
+* [Credits](#credits)
+* [Contributing](#contributing)
+* [License](#license)
+* [Tests](#tests)
+* [Features](#features)
+* [Questions](#questions)
+
+## Installation
+
+
+
+## Usage
+
+
+
+## Video
+
+N/A
+
+## Deployed Site
+
+
+
+## Credits
+
+
+
+## Contributing
+
+N/A
+
+## License
+
+MIT License
+
+Copyright (c) 2024 Tyler-KD
+
+Permission is hereby granted, free of charge, to any person obtaining a copy
+of this software and associated documentation files (the "Software"), to deal
+in the Software without restriction, including without limitation the rights
+to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+copies of the Software, and to permit persons to whom the Software is
+furnished to do so, subject to the following conditions:
+
+The above copyright notice and this permission notice shall be included in all
+copies or substantial portions of the Software.
+
+THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+SOFTWARE.
+
+* (https://choosealicense.com/licenses/mit/)
+
+## Tests
+
+N/A
+
+## Features
+
+
+
+## Questions
+
+If you have any questions, please visit [GitHub/Tyler-KD](https://github.com/Tyler-KD) or submit questions to tyler.kd.knapp@gmail.com.
