+++ app.py
@@ -1,3 +1,7 @@
+# # TODO: Make the user input their own API key
+# # TODO: change it so that the app connects with a different python file that gives a response to the user
+# # TODO: test what files do and don't work for summarization 
+
 from shiny import App, render, ui, reactive
 import os
 import shutil
@@ -41,16 +45,28 @@ app_ui = ui.page_fluid(
             }
         """)
     ),
-    ui.h1("File Upload and Local Storage"),
-    ui.input_file("file_upload", "Upload a file", multiple=True),
-    ui.output_text_verbatim("file_info"), # fix this so that the files uploaded go straight to the list of "uploaded files"
-    ui.markdown("Uploaded files"),
-    ui.output_ui("uploaded_files_list"),
-    ui.input_action_button("delete_button", "Delete Selected Files"),
-    ui.input_action_button("process_button", "Process Selected Files"),
-    ui.input_text("user_question", "Ask a question about the document(s):"),
-    ui.input_action_button("submit_question", "Submit Question"),
-    ui.output_text_verbatim("process_output")
+    ui.page_sidebar(
+        ui.sidebar(
+            ui.accordion(
+                ui.accordion_panel("Uploaded files",
+                    ui.output_ui("uploaded_files_list"),
+                    ui.input_action_button("delete_button", "Delete Selected Files"),
+                    ui.input_action_button("process_button", "Process Selected Files"),
+                ),
+            ),
+            ui.accordion(
+                ui.accordion_panel("Selected files",
+                    ui.output_ui("checked_boxes_list"),
+                ),
+            ),
+        ),
+        ui.h1("File Upload and Local Storage"),
+        ui.input_file("file_upload", "Upload a file", multiple=True),
+        ui.output_text_verbatim("file_info"), # fix this so that the files uploaded go straight to the list of "uploaded files"
+        ui.input_text("user_question", "Ask a question about the document(s):"),
+        ui.input_action_button("submit_question", "Submit Question"),
+        ui.output_text_verbatim("process_output")
+    )
 )
 
 # parent server
@@ -59,6 +75,7 @@ def server(input, output, session):
     delete_trigger = reactive.Value(0)
     process_output_value = reactive.Value("")
     conversation_history = reactive.Value([])
+    
     
     @output
     @render.text
@@ -77,8 +94,6 @@ def server(input, output, session):
             shutil.move(file_path, save_path)
             file_details.append(f"File uploaded: {filename}")
         
-        # Trigger update for the file list UI
-        delete_trigger.set(delete_trigger() + 1)
         return "\n".join(file_details)
     
     @output
@@ -95,6 +110,23 @@ def server(input, output, session):
         ) for file in file_list]
         #  makes an unordered list of the files
         return ui.tags.ul(*checkboxes)
+    
+    @output
+    @render.ui
+    def checked_boxes_list():
+        delete_trigger()  # To make the UI reactive to deletions
+        file_list = os.listdir('uploaded_files')
+        checked_files = []
+        
+        for file in file_list:
+            checkbox_id = f"select_{sanitize_filename(file)}"
+            if input[checkbox_id]():
+                checked_files.append(file)
+        
+        if not checked_files:
+            return ui.p("No files selected.")
+        
+        return ui.tags.ul(*[ui.tags.li(file) for file in checked_files])
 
     @reactive.Effect
     @reactive.event(input.delete_button)
@@ -179,7 +211,7 @@ def server(input, output, session):
                 with open(file_path, 'r') as f:
                     content = f.read()
             document_contents.append(f"File: {os.path.basename(file_path)}\nContent: {content}")
-
+            
         combined_content = "\n\n".join(document_contents)
 
         history = conversation_history.get()
@@ -212,3 +244,103 @@ app = App(app_ui, server)
 
 if __name__ == "__main__":
     app.run()
+
+
+# from shiny import App, render, ui, reactive
+# import os
+# import shutil
+
+# def sanitize_filename(filename):
+#     return "".join(c if c.isalnum() else "_" for c in filename)
+
+# app_ui = ui.page_fluid(
+#     ui.h1("File Upload and Local Storage"),
+#     ui.input_file("file_upload", "Upload a file", multiple=True),
+#     ui.output_text_verbatim("file_info"),
+#     ui.markdown("Uploaded files"),
+#     ui.output_ui("uploaded_files_list"),
+#     ui.markdown("Checked boxes"),
+#     ui.output_ui("checked_boxes_list"),
+#     ui.input_action_button("delete_button", "Delete Selected Files")
+# )
+
+# # Server logic
+# def server(input, output, session):
+#     delete_trigger = reactive.Value(0)
+
+#     @output
+#     @render.text
+#     def file_info():
+#         if input.file_upload() is None:
+#             return "No file uploaded yet."
+        
+#         files = input.file_upload()
+#         file_details = []
+        
+#         for file in files:
+#             filename = file["name"]
+#             file_path = file["datapath"]
+            
+#             # Move the uploaded file to the destination directory
+#             save_path = os.path.join("uploaded_files", filename)
+#             shutil.move(file_path, save_path)
+            
+#             file_details.append(f"File uploaded: {filename}")
+        
+#         return "\n".join(file_details)
+    
+#     @output
+#     @render.ui
+#     def uploaded_files_list():
+#         delete_trigger()  # To make the UI reactive to deletions
+#         file_list = os.listdir('uploaded_files')
+#         if not file_list:
+#             return ui.p("No files uploaded yet.")
+        
+#         checkboxes = [ui.tags.li(
+#             ui.input_checkbox(f"select_{sanitize_filename(file)}", file, value=False),
+#         ) for file in file_list]
+        
+#         return ui.tags.ul(*checkboxes)
+
+#     @output
+#     @render.ui
+#     def checked_boxes_list():
+#         delete_trigger()  # To make the UI reactive to deletions
+#         file_list = os.listdir('uploaded_files')
+#         checked_files = []
+        
+#         for file in file_list:
+#             checkbox_id = f"select_{sanitize_filename(file)}"
+#             if input[checkbox_id]():
+#                 checked_files.append(file)
+        
+#         if not checked_files:
+#             return ui.p("No files selected.")
+        
+#         return ui.tags.ul(*[ui.tags.li(file) for file in checked_files])
+
+#     @reactive.Effect
+#     @reactive.event(input.delete_button)
+#     def delete_selected_files():
+#         file_list = os.listdir('uploaded_files')
+#         for file in file_list:
+#             checkbox_id = f"select_{sanitize_filename(file)}"
+#             if input[checkbox_id]():
+#                 try:
+#                     os.remove(os.path.join('uploaded_files', file))
+#                 except Exception as e:
+#                     print(f"Error deleting {file}: {e}")
+        
+#         # Trigger reactivity to update the file list
+#         delete_trigger.set(delete_trigger() + 1)
+
+# # Create the app
+# app = App(app_ui, server)
+
+# # Ensure the uploaded_files directory exists
+# if not os.path.exists('uploaded_files'):
+#     os.makedirs('uploaded_files')
+
+# if __name__ == "__main__":
+#     app.run()
